[#1375] Revise public API
move KeysResource and Factory to internal package
